# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-PERSON ",
        "description": "Documentação da API-PERSON com Swagger Service.",
        "contact": {
            "name": "Luiz Santos Full Stack Developer ",
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/servidores-efetivos": {
            "get": {
                "tags": [
                    "Servidor Efetivo"
                ],
                "summary": "Lista todos as servidores efetivos",
                "description": "Retorna uma lista de servidores efetivos armazenados no banco de dados.",
                "operationId": "52e4ce496fb4d158b46016deec180806",
                "responses": {
                    "200": {
                        "description": "Lista de servidores efetivos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ServidorEfetivo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum servidor efetivo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não há servidores efetivos cadastrados!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao buscar os servidores efetivos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar os servidores efetivos."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Servidor Efetivo"
                ],
                "summary": "Cria uma nova servidor efetivo",
                "description": "Registra um novo servidor efetivo no banco de dados.",
                "operationId": "db3aca000330b5d6d5334b33a2b8f38c",
                "requestBody": {
                    "description": "Dados necessários para criar um novo servidor efetivo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_id",
                                    "se_matricula"
                                ],
                                "properties": {
                                    "pes_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "se_matricula": {
                                        "type": "string",
                                        "example": "2003456788467"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servidor Efetivo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Servidor Efetivo criado com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServidorEfetivo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar o servidor efetivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar o servidor efetivo."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/servidores-efetivos/{id}": {
            "put": {
                "tags": [
                    "Servidor Efetivo"
                ],
                "summary": "Atualiza um servidor efetivo existente",
                "description": "Atualiza os dados de uma servidor efetivo com base no ID fornecido.",
                "operationId": "6a0d14a7c97f6bc6ea869a121090774d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da servidor efetivo a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para atualização da servidor efetivo",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_id",
                                    "se_matricula"
                                ],
                                "properties": {
                                    "pes_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "se_matricula": {
                                        "type": "string",
                                        "example": "2003456788467"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servidor Efetivo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Servidor Efetivo atualizado com sucesso!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServidorEfetivo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao processar a atualização da servidor efetivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao processar a solicitação."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Servidor Efetivo"
                ],
                "summary": "Exclui um servidor efetivo",
                "description": "Exclui um servidor efetivo do banco de dados com base no ID fornecido.",
                "operationId": "7769253db1ca2a0f2705f4d8cb7b5938",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da servidor efetivo a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pessoa excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pessoa excluído com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao excluir o servidor efetivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao excluir o servidor efetivo."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ServidorEfetivo": {
                "title": "Servidor Efetivo",
                "description": "Modelo de Servidor Efetivo",
                "required": [
                    "pes_id",
                    "se_matricula"
                ],
                "properties": {
                    "pes_id": {
                        "description": "ID da pessoa",
                        "type": "integer",
                        "example": "2"
                    },
                    "se_matricula": {
                        "description": "Matricula do Servidor",
                        "type": "string",
                        "format": "date-time",
                        "example": "2003456788467"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}